requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
umap-learn
pgmpy
networkx
scipy
joblib
openpyxl